cleaned up console logs and added comments

+++ src/components/VolunteerSchedule/Manage/Manage.jsx
@@ -9,7 +9,9 @@ import Button from '@material-ui/core/Button';
 import './Manage.css';
 
 class Manage extends Component {
+
     state = {}
+
     componentDidMount() {
         axios.get('/api/volunteer-admin/shifts')
             .then(response => {
@@ -21,6 +23,7 @@ class Manage extends Component {
                 console.log(error)
             })
     }
+
     findUniqueShifts = (shifts) => {
         let allShiftTimes = [];
         // pushes all shifts date/time into array
@@ -36,13 +39,14 @@ class Manage extends Component {
         // console.log(arrayOfTimesToReturn)
         return arrayOfTimesToReturn
     }
+
     processDataToSend = (data) => {
         let dataToSend = [];
         data.forEach(row => {
             dataToSend.push({
                 department: row.department,
                 role: row.role,
-                okForWalkUps: row.ok_for_walk_ups, 
+                okForWalkUps: row.ok_for_walk_ups,
                 allShifts: row.shifts,
                 uniqueShifts: this.findUniqueShifts(row.shifts),
                 RoleID: row.RoleID
@@ -53,13 +57,13 @@ class Manage extends Component {
     }
 
     deleteAll = () => {
-            swal({
-                title: `Are you sure? This will delete all shifts and associated volunteer hours.`,
-                text: `After deletion you will be able to create a new schedule by uploading a CSV file. If you are resetting after an event, you may want to create a separate record of total volunteer hours. These can be found under Volunteer Contacts.`,
-                icon: "warning",
-                buttons: true,
-                dangerMode: true,
-            }).then(function(willDelete) {
+        swal({
+            title: `Are you sure? This will delete all shifts and associated volunteer hours.`,
+            text: `After deletion you will be able to create a new schedule by uploading a CSV file. If you are resetting after an event, you may want to create a separate record of total volunteer hours. These can be found under Volunteer Contacts.`,
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
+        }).then(function (willDelete) {
             if (willDelete) {
                 axios.delete(`/api/volunteer-admin/delete-schedule/`)
                     .then(response => {
@@ -72,24 +76,24 @@ class Manage extends Component {
                     }).catch(error => {
                         console.log(error)
                     })
-                } else {
+            } else {
                 swal("Your schedule is safe!");
-                }
-            });
-        }
+            }
+        });
+    }
 
-        previousPage = () => {
-            this.props.history.push('/volunteer-schedule')
-        }
+    previousPage = () => {
+        this.props.history.push('/volunteer-schedule')
+    }
 
     render() {
         return (
             <div className="Manage">
                 <div className="back-button">
-                <Button variant="contained" onClick={this.previousPage}>Back</Button>
+                    <Button variant="contained" onClick={this.previousPage}>Back</Button>
                 </div>
                 <h1>Manage Volunteer Schedule</h1>
-                {this.state.data ? 
+                {this.state.data ?
                     <div>
                         {this.state.data.map(row => (
                             <Accordion
@@ -99,16 +103,16 @@ class Manage extends Component {
                         ))}
                     </div>
                     : 'loading…'}
-                    <div className="deleteAll">
-                        <Button
+                <div className="deleteAll">
+                    <Button
                         variant="contained"
                         color="secondary"
                         onClick={this.deleteAll}
                         startIcon={<DeleteIcon />}
-                        >
+                    >
                         Delete Entire Schedule
                         </Button>
-                    </div>
+                </div>
             </div >
         );
     }
